Move mapping class to Resources' sub namespace and make Resources to an own namespace
