+++ Controllers/OrdersController.cs
@@ -2,8 +2,8 @@ using System;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using Warehouse.Controllers.Resources.Requests;
-using Warehouse.Controllers.Resources.Responses;
+using Warehouse.Resources.Requests;
+using Warehouse.Resources.Responses;
 using Warehouse.Core;
 using Warehouse.Core.Facades;
 using Warehouse.Core.Models;
