+++ Controllers/SuppliesController.cs
@@ -4,8 +4,8 @@ using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using Warehouse.Core.Models;
 using Warehouse.Core.Repositories;
-using Warehouse.Controllers.Resources.Requests;
-using Warehouse.Controllers.Resources.Responses;
+using Warehouse.Resources.Requests;
+using Warehouse.Resources.Responses;
 using Warehouse.Core;
 using Warehouse.Core.Facades;
