+++ Controllers/TechnicianOrdersController.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using Warehouse.Controllers.Resources.Responses;
+using Warehouse.Resources.Responses;
 using Warehouse.Core.Models;
 using Warehouse.Core.Repositories;
