+++ Resources/Requests/SaveMerchantResource.cs
@@ -1,22 +1,20 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Warehouse.Controllers.Resources.Requests
+namespace Warehouse.Resources.Requests
 {
-    public class SaveTechnicianResource
+    public class SaveMerchantResource
     {
         [Required]
         [StringLength(255)]
-        public string FirstName { get; set; }
-
+        public string Name { get; set; }
+        
         [Required]
         [StringLength(255)]
-        public string LastName { get; set; }
-
-        [Required]
+        public string City { get; set; }
+        
         [StringLength(255)]
         public string Email { get; set; }
-
-        [Required]
+        
         [StringLength(25)]
         public string Phone { get; set; }
     }
