+++ Resources/Requests/SaveTechnicianResource.cs
@@ -1,20 +1,22 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Warehouse.Controllers.Resources.Requests
+namespace Warehouse.Resources.Requests
 {
-    public class SaveMerchantResource
+    public class SaveTechnicianResource
     {
         [Required]
         [StringLength(255)]
-        public string Name { get; set; }
-        
+        public string FirstName { get; set; }
+
         [Required]
         [StringLength(255)]
-        public string City { get; set; }
-        
+        public string LastName { get; set; }
+
+        [Required]
         [StringLength(255)]
         public string Email { get; set; }
-        
+
+        [Required]
         [StringLength(25)]
         public string Phone { get; set; }
     }
