+++ Resources/Responses/MerchantKeyValuePairWithProductKeyValuePairsResource.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 
-namespace Warehouse.Controllers.Resources.Responses
+namespace Warehouse.Resources.Responses
 {
     public class MerchantKeyValuePairWithProductKeyValuePairsResource : KeyValuePairResource
     {
