+++ Resources/Responses/PhotoResource.cs
@@ -1,4 +1,4 @@
-namespace Warehouse.Controllers.Resources.Responses
+namespace Warehouse.Resources.Responses
 {
     public class PhotoResource
     {
