+++ Resources/Responses/SupplyEntryResource.cs
@@ -1,9 +1,12 @@
-namespace Warehouse.Controllers.Resources.Responses
+using System;
+
+namespace Warehouse.Resources.Responses
 {
-    public class SupplyResource
+    public class SupplyEntryResource
     {
         public int Id { get; set; }
         public ProductResource Product { get; set; }
         public int Quantity { get; set; }
+        public DateTime CreatedAt { get; set; }
     }
 }
