+++ Resources/Responses/SupplyResource.cs
@@ -1,0 +1,9 @@
+namespace Warehouse.Resources.Responses
+{
+    public class SupplyResource
+    {
+        public int Id { get; set; }
+        public ProductResource Product { get; set; }
+        public int Quantity { get; set; }
+    }
+}
